change pocketModel.Balance from int64 to float 64

+++ pocket/pocket.go
@@ -3,11 +3,11 @@ package pocket
 import "database/sql"
 
 type PocketModel struct {
-	ID       int    `json:"id"`
-	Name     string `json:"name"`
-	Category string `json:"category"`
-	Currency string `json:"currency"`
-	Balance  int64  `json:"balance"`
+	ID       int     `json:"id"`
+	Name     string  `json:"name"`
+	Category string  `json:"category"`
+	Currency string  `json:"currency"`
+	Balance  float64 `json:"balance"`
 }
 
 type handler struct {
